+++ R/long_masterpull.R
@@ -1,7 +1,7 @@
 #' A function to import a longitudinal version of EdBuild's master dataset
 #'
 #' This function imports a longitudinal dataset of EdBuild's master data for the
-#' years 2013- 2017. The master dataset is a compilation of national level
+#' years 2013- 2018. The master dataset is a compilation of national level
 #' school district data from the US Census Annual Survey of School System
 #' Finances (F33); US Census Small Area Income and Poverty Estimates (SAIPE);
 #' National Center for Education Statistics (NCES) Common Core of Data (CCD);
@@ -13,7 +13,7 @@
 #' @return A dataframe where each observation is a school district.
 #' @seealso \code{\link{master_codebook}}, \code{\link{masterpull}}
 #' @export
-#' @format A data frame with 92,636 observations and 41 variables. To view
+#' @format A data frame with 111,264 observations and 43 variables. To view
 #'   descriptions of variable names and sources for each use
 #'   \code{master_codebook()}
 #' @examples
@@ -27,6 +27,7 @@ long_masterpull = function() {
       url_15 = "https://s3.amazonaws.com/data.edbuild.org/public/Processed+Data/Master/2015/full_data_15.csv"
       url_16 = "https://s3.amazonaws.com/data.edbuild.org/public/Processed+Data/Master/2016/full_data_16.csv"
       url_17 = "https://s3.amazonaws.com/data.edbuild.org/public/Processed+Data/Master/2017/full_data_17.csv"
+      url_18 = "https://s3.amazonaws.com/data.edbuild.org/public/Processed+Data/Master/2018/full_data_18.csv"
 
       master_13 <- read.csv(file = url_13, stringsAsFactors = FALSE) %>%
         dplyr::mutate(CONUM = as.character(CONUM),
@@ -58,8 +59,14 @@ long_masterpull = function() {
                       FRL_rate = dFRL/dEnroll_district,
                       year = "2017")
 
-      long_master <- dplyr::bind_rows(master_13, master_14, master_15, master_16, master_17) %>%
-        dplyr::select(-Region, -X, -COLIn) %>%
+      master_18 <- read.csv(file = url_18, stringsAsFactors = FALSE) %>%
+        dplyr::mutate(CONUM = as.character(CONUM),
+                      STATE_FIPS = as.character(STATE_FIPS),
+                      FRL_rate = dFRL/dEnroll_district,
+                      year = "2018")
+
+      long_master <- dplyr::bind_rows(master_13, master_14, master_15, master_16, master_17, master_18) %>%
+        dplyr::select(-X) %>%
         dplyr::select(NCESID, year, everything())
 
     return(long_master)